+++ changelog.md
@@ -14,6 +14,7 @@ Codion Change Log
 - ControlKey.name() added.
 ### is.codion.swing.framework.ui
 - EntityTablePanel bug fixed, summary panel toggle control displayed only if a summary panel is available.
+- EntityApplicationPanel.createMainMenuControls() and related methods now Optional based.
 
 ## 0.18.2
 ### is.codion.swing.common.model